block_committer tests now integration tests
