+++ .sqlx/query-2e9f0e15644765d9276402081dd759b5608af2c2fe2fcc4e79ef7d8184ceece9.json
@@ -1,0 +1,58 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            WITH\n\n            -- Delete from l1_blob_transaction\n            deleted_blob_transactions AS (\n                DELETE FROM l1_blob_transaction\n                WHERE created_at < $1\n                RETURNING id\n            ),\n\n            -- Delete from l1_transaction_fragments\n            deleted_transaction_fragments AS (\n                DELETE FROM l1_transaction_fragments\n                WHERE transaction_id IN (SELECT id FROM deleted_blob_transactions)\n                RETURNING transaction_id\n            ),\n\n            -- Build updated_transaction_fragments that represent the state after deletions\n            updated_transaction_fragments AS (\n                SELECT fragment_id FROM l1_transaction_fragments\n                WHERE transaction_id NOT IN (SELECT transaction_id FROM deleted_transaction_fragments)\n            ),\n\n            -- Delete unreferenced fragments\n            deleted_fragments AS (\n                DELETE FROM l1_fragments\n                WHERE id NOT IN (SELECT fragment_id FROM updated_transaction_fragments)\n                RETURNING id\n            ),\n\n            -- Step 4: Build updated_fragments that represent the state of after deletions\n            updated_fragments AS (\n                SELECT bundle_id FROM l1_fragments\n                WHERE id NOT IN (SELECT id FROM deleted_fragments)\n            ),\n\n            -- Delete unreferenced bundles and collect start and end heights\n            deleted_bundles AS (\n                DELETE FROM bundles\n                WHERE id NOT IN (SELECT bundle_id FROM updated_fragments)\n                RETURNING start_height, end_height\n            ),\n\n            -- Delete corresponding fuel_blocks entries\n            deleted_fuel_blocks AS (\n                DELETE FROM fuel_blocks\n                WHERE EXISTS (\n                    SELECT 1 FROM deleted_bundles\n                    WHERE fuel_blocks.height BETWEEN deleted_bundles.start_height AND deleted_bundles.end_height\n                )\n                RETURNING height\n            ),\n\n            -- Delete from l1_transaction\n            deleted_transactions AS (\n                DELETE FROM l1_transaction\n                WHERE created_at < $1\n                RETURNING id\n            ),\n\n            -- Build updated_contract transaction that represent the state after deletions\n            updated_transactions AS (\n                SELECT submission_id FROM l1_transaction\n                WHERE id NOT IN (SELECT id FROM deleted_transactions)\n            ),\n\n            -- Delete from l1_fuel_block_submission\n            deleted_submissions AS (\n                DELETE FROM l1_fuel_block_submission\n                WHERE id NOT IN (SELECT submission_id FROM updated_transactions)\n                RETURNING id\n            )\n\n            SELECT\n                (SELECT COUNT(*) FROM deleted_blob_transactions) AS deleted_blob_transactions,\n                (SELECT COUNT(*) FROM deleted_transaction_fragments) AS deleted_transaction_fragments,\n                (SELECT COUNT(*) FROM deleted_fragments) AS deleted_fragments,\n                (SELECT COUNT(*) FROM deleted_bundles) AS deleted_bundles,\n                (SELECT COUNT(*) FROM deleted_fuel_blocks) AS deleted_fuel_blocks,\n                (SELECT COUNT(*) FROM deleted_transactions) AS deleted_contract_transactions,\n                (SELECT COUNT(*) FROM deleted_submissions) AS deleted_contract_submissions;\n            ",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "deleted_blob_transactions",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 1,
+        "name": "deleted_transaction_fragments",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 2,
+        "name": "deleted_fragments",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 3,
+        "name": "deleted_bundles",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 4,
+        "name": "deleted_fuel_blocks",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 5,
+        "name": "deleted_contract_transactions",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 6,
+        "name": "deleted_contract_submissions",
+        "type_info": "Int8"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Timestamptz"
+      ]
+    },
+    "nullable": [
+      null,
+      null,
+      null,
+      null,
+      null,
+      null,
+      null
+    ]
+  },
+  "hash": "2e9f0e15644765d9276402081dd759b5608af2c2fe2fcc4e79ef7d8184ceece9"
+}
