+++ packages/test-helpers/Cargo.toml
@@ -14,6 +14,7 @@ clock = { workspace = true, features = ["test-helpers"] }
 delegate = { workspace = true }
 eth = { workspace = true, features = ["test-helpers"] }
 fuel-block-committer-encoding = { workspace = true }
+fuel-crypto = { workspace = true, features = ["random"] }
 futures = { workspace = true }
 itertools = { workspace = true, features = ["use_alloc"] }
 metrics = { workspace = true }
@@ -22,3 +23,4 @@ ports = { workspace = true, features = ["full", "test-helpers"] }
 rand = { workspace = true, features = ["small_rng"] }
 services = { workspace = true, features = ["test-helpers"] }
 storage = { workspace = true, features = ["test-helpers"] }
+tai64 = { workspace = true }
